+++ src/constants/index.js
@@ -3,3 +3,5 @@ export { COMMODITIES } from './commodities.js';
 export { GOV_TYPES, TECH_LEVELS, SIZES, SYSTEM_NAMES } from './world.js';
 export { EVENT_TEMPLATES, CONTRACT_NAMES } from './events.js';
 export { ELITE_CAPTAINS, MERCENARY_POOL } from './mercenaries.js';
+export { ALIEN_SHIPS, ALIEN_DISRUPTOR, ALIEN_GADGETS, ALIEN_ARTIFACT,
+         OCCUPATION_STAGES, OCCUPATION_SPEED } from './aliens.js';